test: Move helpers to a different file
Sucks having to scroll through helpers to get to the tests.
